requirements.txt
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.4.0
Django==2.2.6
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.9.0
entrypoints==0.3
flake8==3.7.8
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycodestyle==2.5.0
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.2
pytz==2019.3
PyYAML==5.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.3
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
